portfolio-tracker/requirements.txt
# Web Framework
streamlit==1.40.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Data & Finance
pandas==2.2.3
numpy==1.26.4
yfinance==0.2.50

# Visualization
plotly==5.24.1

# Utilities
python-dotenv==1.0.1
bcrypt==4.2.1

# Development
pytest==8.3.4
black==24.10.0